predictor needs some more space internally

+++ include/adpcm.hpp
@@ -15,7 +15,7 @@ namespace Csdr {
             int16_t getPredictor();
         private:
             int16_t index = 0;         // Index into step size table
-            int16_t previousValue = 0; // Most recent sample value
+            int32_t previousValue = 0; // Most recent sample value
             const int _stepSizeTable[89] = {
                     7, 8, 9, 10, 11, 12, 13, 14, 16, 17, 19, 21, 23, 25, 28, 31, 34,
                     37, 41, 45, 50, 55, 60, 66, 73, 80, 88, 97, 107, 118, 130, 143,
